requirements.txt
# Core packages
yfinance
newsapi-python

# New API clients
alpha-vantage  # For Alpha Vantage API (alternative: requests)
fredapi       # For FRED economic data (alternative: requests)

# Hugging Face Transformers for sentiment analysis
transformers
torch

# Web app UI
streamlit

# Plotting (optional for visuals)
matplotlib
seaborn
plotly

# LangChain (future use)
langchain

# For SQLite & Postgres support
sqlite-utils
psycopg2-binary  # for QuestDB PostgreSQL compatibility

# Other utils
numpy
requests
pandas
praw
python-dotenv
emoji
langdetect
tweepy